cast to remove warning.
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@22777 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/bzflag/playing.cxx
@@ -1565,7 +1565,7 @@ static void loadCachedWorld()
   // get the world size
   cachedWorld->seekg(0, std::ios::end);
   std::streampos size = cachedWorld->tellg();
-  unsigned long charSize = std::streamoff(size);
+  unsigned long charSize = (unsigned long)std::streamoff(size);
 
   // load the cached world
   cachedWorld->seekg(0);
